requirements.txt
fastapi
uvicorn
sentence-transformers
numpy
pydantic
mysql-connector-python
thefuzz
python-Levenshtein
python-dotenv
